+++ lib/contextual_logger.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require 'logger' # required first to get ::Logger defined for ActiveSupport 7.0
 require 'active_support'
 require 'active_support/core_ext/module/delegation'
 require 'json'
@@ -11,12 +12,12 @@ require_relative './contextual_logger/global_context_lock_message'
 module ContextualLogger
   LOG_LEVEL_NAMES_TO_SEVERITY =
   {
-    debug:  Logger::Severity::DEBUG,
-    info:   Logger::Severity::INFO,
-    warn:   Logger::Severity::WARN,
-    error:  Logger::Severity::ERROR,
-    fatal:  Logger::Severity::FATAL,
-    unknown: Logger::Severity::UNKNOWN
+    debug:  ::Logger::Severity::DEBUG,
+    info:   ::Logger::Severity::INFO,
+    warn:   ::Logger::Severity::WARN,
+    error:  ::Logger::Severity::ERROR,
+    fatal:  ::Logger::Severity::FATAL,
+    unknown: ::Logger::Severity::UNKNOWN
   }.freeze
 
   class << self