+++ lib/contextual_logger/logger_with_context.rb
@@ -40,6 +40,12 @@ module ContextualLogger
       @override_level = (ContextualLogger.normalize_log_level(override_level) if override_level)
     end
 
+    ::ContextualLogger::LOG_LEVEL_NAMES_TO_SEVERITY.except(:unknown).each do |severity, log_level|
+      define_method("#{severity}?") do
+        log_level_enabled?(log_level)
+      end
+    end
+
     def write_entry_to_log(severity, timestamp, progname, message, context:)
       merged_context =
         if context.any?
@@ -54,16 +60,13 @@ module ContextualLogger
     private
 
     def normalize_context(context)
-      if warn_on_string_keys(context)
-        context.deep_symbolize_keys
-      else
-        context
-      end
+      raise_on_string_keys(context)
+      context
     end
 
-    def warn_on_string_keys(context)
+    def raise_on_string_keys(context)
       if deep_key_has_string?(context)
-        ActiveSupport::Deprecation.warn('Context keys must use symbols not strings. This will be asserted as of contextual_logger v1.0.0')
+        raise ArgumentError, "context keys must use symbols not strings: #{context.inspect}"
       end
     end