+++ spec/lib/contextual_logger/logger_with_context_spec.rb
@@ -126,24 +126,60 @@ describe ContextualLogger::LoggerWithContext do
       end
     end
 
+    describe "log level predicates" do
+      it "delegates to DEBUG level" do
+        subject.level = Logger::Severity::DEBUG
+        expect(subject.debug?).to eq(true)
+        expect(subject.info?).to eq(true)
+        expect(subject.warn?).to eq(true)
+        expect(subject.error?).to eq(true)
+        expect(subject.fatal?).to eq(true)
+      end
+
+      it "delegates to INFO level" do
+        subject.level = Logger::Severity::INFO
+        expect(subject.debug?).to eq(false)
+        expect(subject.info?).to eq(true)
+        expect(subject.warn?).to eq(true)
+        expect(subject.error?).to eq(true)
+        expect(subject.fatal?).to eq(true)
+      end
+
+      it "delegates to WARN level" do
+        subject.level = nil
+        base_logger.level = Logger::Severity::WARN
+        expect(subject.debug?).to eq(false)
+        expect(subject.info?).to eq(false)
+        expect(subject.warn?).to eq(true)
+        expect(subject.error?).to eq(true)
+        expect(subject.fatal?).to eq(true)
+      end
+
+      it "delegates to ERROR level" do
+        subject.level = Logger::Severity::ERROR
+        expect(subject.debug?).to eq(false)
+        expect(subject.info?).to eq(false)
+        expect(subject.warn?).to eq(false)
+        expect(subject.error?).to eq(true)
+        expect(subject.fatal?).to eq(true)
+      end
+
+      it "delegates to FATAL level" do
+        subject.level = Logger::Severity::FATAL
+        expect(subject.debug?).to eq(false)
+        expect(subject.info?).to eq(false)
+        expect(subject.warn?).to eq(false)
+        expect(subject.error?).to eq(false)
+        expect(subject.fatal?).to eq(true)
+      end
+    end
+
     context "when string passed as context key" do
-      it "returns context with a symbol key" do
+      it "raises ArgumentError" do
         context_with_string_key = { "log_source" => "redis_client" }
-        string_context = ContextualLogger::LoggerWithContext.new(base_logger, context_with_string_key)
-        expect(string_context.context).to eq(log_source: "redis_client")
-      end
-
-      it "returns a deep context with symbol key" do
-        context_with_string_key_levels = { log_source: { level1: { level2: { "level3" => "redis_client" } } } }
-        string_context = ContextualLogger::LoggerWithContext.new(base_logger, context_with_string_key_levels)
-        expect(string_context.context)
-          .to eq({ log_source: { level1: { level2: { level3: "redis_client" } } } })
-      end
-
-      it "should return a deprecation warning" do
-        context_with_string_key = { "log_source" => "redis_client" }
-        expect { ContextualLogger::LoggerWithContext.new(base_logger, context_with_string_key) }
-          .to output(/DEPRECATION WARNING: Context keys must use symbols not strings/).to_stderr
+        expect do
+          ContextualLogger::LoggerWithContext.new(base_logger, context_with_string_key)
+        end.to raise_exception(ArgumentError, /context keys must use symbols not strings: \{"log_source" ?=> ?"redis_client"\}/)
       end
     end