+++ CHANGELOG.md
@@ -4,6 +4,12 @@ Inspired by [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 Note: this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0] - 2025-09-12
+### Added
+- Added LoggerWithContext support for `debug?`, `info?`, `warn?`, `error?`, `fatal?`.
+### Changed
+- Raise ArgumentError if context keys are not symbols. (Was documented to start in v1.0.0 but was not implemented until now.)
+
 ## [1.4.0] - 2024-07-10
 ### Added
 - Added support for `activesupport` 7.1 by providing a mixin for extending Broadcast loggers.